eval/requirements.txt
tau2-bench @ git+https://github.com/sierra-research/tau2-bench.git
openai==1.51.0
anthropic==0.34.2
langfuse==2.51.3
scipy==1.14.1
numpy==2.1.2
pandas==2.2.3
python-dotenv==1.0.1
structlog==24.4.0
tqdm==4.66.5
